Major UPDATE: Read  Description
Lots of UI changes, added the ability to see the igredients and directions for the recipe with user clicking the enlarge button. Improved the code with keeping count of the recipes added. Simplified the code and seperated their duties into separate functions all working together, added the date for each object created.

For example, to get the enlarge see full details button to work;

1.  Inside the UI module, I created getFullRecipe to see if  a DOMTokenList contained the value of id, if yess, we store it into a value and split it by the hyphen.

2. The value is returned and  is used inside the internal module; this function checks to see if the ids match, then returns  teh corresponding object with that id.

3. That returned value from (2) is passed into another function inside the UI module, it will take the HTML textcontent and store the object's properties.