Project complete according to requirements but More comments/readme doc in the future: Add/delete/edit finished
What needs to be done
Media  queries, better UX design, add images, the this week/day/month functions need to be set up and local storage needs to be set up. I'll  return to this project sometime to offfically finish it but I have a long ways to go and need to cover more material.